+++ Tests/prlctlTests/VMTests.swift
@@ -108,33 +108,32 @@ final class VMTests: XCTestCase {
 
     func testThatRunningVMCanBeStopped() throws {
         let runner = TestCommandRunner()
-        try RunningVM(uuid: "machine-uuid", name: "machine-name", ipAddress: "127.0.0.1").shutdown(runner: runner)
+        try RunningVM.testCase.shutdown(runner: runner)
         XCTAssertEqual(runner.command, "prlctl stop machine-uuid")
     }
 
     func testThatRunningVMCanBeStoppedImmediately() throws {
         let runner = TestCommandRunner()
-        try RunningVM(uuid: "machine-uuid", name: "machine-name", ipAddress: "127.0.0.1").shutdown(immediately: true, runner: runner)
+        try RunningVM.testCase.shutdown(immediately: true, runner: runner)
         XCTAssertEqual(runner.command, "prlctl stop machine-uuid --fast")
     }
 
     func testThatRunningVMCanPerformCommandsAsRoot() throws {
         let runner = TestCommandRunner()
-        try RunningVM(uuid: "machine-uuid", name: "machine-name", ipAddress: "127.0.0.1", runner: runner).runCommand("ls")
+        try RunningVM.testCase.runCommand("ls", runner: runner)
         XCTAssertEqual(runner.command, "prlctl exec machine-uuid ls")
     }
 
     func testThatRunningVMCanPerformCommandsAsAnyUser() throws {
         let runner = TestCommandRunner()
-        try RunningVM(uuid: "machine-uuid", name: "machine-name", ipAddress: "127.0.0.1", runner: runner).runCommand("ls", as: User(named: "user"))
+        try RunningVM.testCase.runCommand("ls", as: User(named: "user"), runner: runner)
         XCTAssertEqual(runner.command, "prlctl exec machine-uuid su - 'user' -c 'ls'")
     }
 
     func testThatRunningVMCanPerformCommandsWithEscapedUsername() throws {
         let runner = TestCommandRunner()
-        try RunningVM(uuid: "machine-uuid", name: "machine-name", ipAddress: "127.0.0.1", runner: runner).runCommand("ls", as: User(named: "my user"))
+        try RunningVM.testCase.runCommand("ls", as: User(named: "my user"), runner: runner)
         XCTAssertEqual(runner.command, "prlctl exec machine-uuid su - 'my user' -c 'ls'")
-
     }
 
     func testThatVMDetailsCanBeParsed() throws {
@@ -144,14 +143,14 @@ final class VMTests: XCTestCase {
     }
 
     func testThatStartedVMWithDetailsCanBeParsed() throws {
-        let codableVM = try XCTUnwrap(JSONDecoder().decode(CodableVM.self, from: getJSONDataForResource(named: "running-vm-with-ip")))
-        let vmDetails = try XCTUnwrap(JSONDecoder().decode(VMDetails.self, from: getJSONDataForResource(named: "running-vm-with-ip-details")))
+        let codableVM = try XCTUnwrap(codableVMResource(named: "running-vm-with-ip"))
+        let vmDetails = try XCTUnwrap(vmDetailsResource(named: "running-vm-with-ip-details"))
         let vm = try XCTUnwrap(VM(vm: codableVM, details: vmDetails))
         XCTAssertEqual(VMStatus.running, vm.status)
     }
 
     func testThatSnapshotListCanBeParsed() throws {
-        let vmList = try XCTUnwrap((JSONDecoder().decode(CodableVMSnapshotList.self, from: getJSONDataForResource(named: "vm-snapshot-list"))))
+        let vmList = try XCTUnwrap((codableVMListResource(named: "vm-snapshot-list")))
         XCTAssertEqual(vmList.count, 1)
     }
 
@@ -186,173 +185,15 @@ final class VMTests: XCTestCase {
     func testThatCleanWorks() throws {
         let runner = TestCommandRunner(response: getJSONResource(named: "vm-snapshot-list"))
         try StoppedVM(uuid: "machine-uuid", name: "machine-name").clean(runner: runner)
-        XCTAssertEqual(runner.commands.last, "prlctl snapshot-delete machine-uuid -i {64d481bb-ce04-45b1-8328-49e4e4c43ddf}")
-    }
-
-    func testThatSetCPUCountWorks() throws {
-        let runner = TestCommandRunner()
-        let vm = StoppedVM(uuid: "machine-uuid", name: "machine-name")
-        try vm.set(.cpuCount(24), runner: runner)
-        XCTAssertEqual(runner.command, "prlctl set machine-uuid --cpus 24")
-    }
-
-    func testThatSetMemorySizeWorks() throws {
-        let runner = TestCommandRunner()
-        let vm = StoppedVM(uuid: "machine-uuid", name: "machine-name")
-        try vm.set(.memorySize(8192), runner: runner)
-        XCTAssertEqual(runner.command, "prlctl set machine-uuid --memsize 8192")
-    }
-
-    func testThatSetHypervisorTypeToParallelsWorks() throws {
-        let runner = TestCommandRunner()
-        let vm = StoppedVM(uuid: "machine-uuid", name: "machine-name")
-        try vm.set(.hypervisorType(.parallels), runner: runner)
-        XCTAssertEqual(runner.command, "prlctl set machine-uuid --hypervisor-type parallels")
-    }
-
-    func testThatSetHypervisorTypeToAppleWorks() throws {
-        let runner = TestCommandRunner()
-        let vm = StoppedVM(uuid: "machine-uuid", name: "machine-name")
-        try vm.set(.hypervisorType(.apple), runner: runner)
-        XCTAssertEqual(runner.command, "prlctl set machine-uuid --hypervisor-type apple")
-    }
-
-    func testThatSetNetworkInterfaceTypeToSharedWorks() throws {
-        let runner = TestCommandRunner()
-        let vm = StoppedVM(uuid: "machine-uuid", name: "machine-name")
-        try vm.set(.networkType(.shared), runner: runner)
-        XCTAssertEqual(runner.command, "prlctl set machine-uuid --device-set net0 --type shared")
-    }
-
-    func testThatSetNetworkInterfaceTypeToBridgedWorks() throws {
-        let runner = TestCommandRunner()
-        let vm = StoppedVM(uuid: "machine-uuid", name: "machine-name")
-        try vm.set(.networkType(.bridged), runner: runner)
-        XCTAssertEqual(runner.command, "prlctl set machine-uuid --device-set net0 --type bridged")
-    }
-
-    func testThatSetNetworkInterfaceTypeToHostOnlyWorks() throws {
-        let runner = TestCommandRunner()
-        let vm = StoppedVM(uuid: "machine-uuid", name: "machine-name")
-        try vm.set(.networkType(.hostOnly), runner: runner)
-        XCTAssertEqual(runner.command, "prlctl set machine-uuid --device-set net0 --type host-only")
-    }
-
-    func testThatSetSmartMountToOnWorks() throws {
-        let runner = TestCommandRunner()
-        let vm = StoppedVM(uuid: "machine-uuid", name: "machine-name")
-        try vm.set(.smartMount(.on), runner: runner)
-        XCTAssertEqual(runner.command, "prlctl set machine-uuid --smart-mount on")
-    }
-
-    func testThatSetSmartMountToOffWorks() throws {
-        let runner = TestCommandRunner()
-        let vm = StoppedVM(uuid: "machine-uuid", name: "machine-name")
-        try vm.set(.smartMount(.off), runner: runner)
-        XCTAssertEqual(runner.command, "prlctl set machine-uuid --smart-mount off")
-    }
-
-    func testThatSetSharedClipboardToOnWorks() throws {
-        let runner = TestCommandRunner()
-        let vm = StoppedVM(uuid: "machine-uuid", name: "machine-name")
-        try vm.set(.sharedClipboard(.on), runner: runner)
-        XCTAssertEqual(runner.command, "prlctl set machine-uuid --shared-clipboard on")
-    }
-
-    func testThatSetSharedClipboardToOffWorks() throws {
-        let runner = TestCommandRunner()
-        let vm = StoppedVM(uuid: "machine-uuid", name: "machine-name")
-        try vm.set(.sharedClipboard(.off), runner: runner)
-        XCTAssertEqual(runner.command, "prlctl set machine-uuid --shared-clipboard off")
-    }
-
-    func testThatSetSharedCloudToOnWorks() throws {
-        let runner = TestCommandRunner()
-        let vm = StoppedVM(uuid: "machine-uuid", name: "machine-name")
-        try vm.set(.sharedCloud(.on), runner: runner)
-        XCTAssertEqual(runner.command, "prlctl set machine-uuid --shared-cloud on")
-    }
-
-    func testThatSetSharedCloudToOffWorks() throws {
-        let runner = TestCommandRunner()
-        let vm = StoppedVM(uuid: "machine-uuid", name: "machine-name")
-        try vm.set(.sharedCloud(.off), runner: runner)
-        XCTAssertEqual(runner.command, "prlctl set machine-uuid --shared-cloud off")
-    }
-
-    func testThatSetSharedProfileToOnWorks() throws {
-        let runner = TestCommandRunner()
-        let vm = StoppedVM(uuid: "machine-uuid", name: "machine-name")
-        try vm.set(.sharedProfile(.on), runner: runner)
-        XCTAssertEqual(runner.command, "prlctl set machine-uuid --shared-profile on")
-    }
-
-    func testThatSetSharedProfileToOffWorks() throws {
-        let runner = TestCommandRunner()
-        let vm = StoppedVM(uuid: "machine-uuid", name: "machine-name")
-        try vm.set(.sharedProfile(.off), runner: runner)
-        XCTAssertEqual(runner.command, "prlctl set machine-uuid --shared-profile off")
-    }
-
-    func testThatSetSharedCameraToOnWorks() throws {
-        let runner = TestCommandRunner()
-        let vm = StoppedVM(uuid: "machine-uuid", name: "machine-name")
-        try vm.set(.sharedCamera(.on), runner: runner)
-        XCTAssertEqual(runner.command, "prlctl set machine-uuid --auto-share-camera on")
-    }
-
-    func testThatSetSharedCameraToOffWorks() throws {
-        let runner = TestCommandRunner()
-        let vm = StoppedVM(uuid: "machine-uuid", name: "machine-name")
-        try vm.set(.sharedCamera(.off), runner: runner)
-        XCTAssertEqual(runner.command, "prlctl set machine-uuid --auto-share-camera off")
-    }
-
-    func testThatSetSharedBluetoothToOnWorks() throws {
-        let runner = TestCommandRunner()
-        let vm = StoppedVM(uuid: "machine-uuid", name: "machine-name")
-        try vm.set(.sharedBluetooth(.on), runner: runner)
-        XCTAssertEqual(runner.command, "prlctl set machine-uuid --auto-share-bluetooth on")
-    }
-
-    func testThatSetSharedBluetoothToOffWorks() throws {
-        let runner = TestCommandRunner()
-        let vm = StoppedVM(uuid: "machine-uuid", name: "machine-name")
-        try vm.set(.sharedBluetooth(.off), runner: runner)
-        XCTAssertEqual(runner.command, "prlctl set machine-uuid --auto-share-bluetooth off")
-    }
-
-    func testThatSetSharedSmartcardToOnWorks() throws {
-        let runner = TestCommandRunner()
-        let vm = StoppedVM(uuid: "machine-uuid", name: "machine-name")
-        try vm.set(.sharedSmartcard(.on), runner: runner)
-        XCTAssertEqual(runner.command, "prlctl set machine-uuid --auto-share-smart-card on")
-    }
-
-    func testThatSetSharedSmartcardToOffWorks() throws {
-        let runner = TestCommandRunner()
-        let vm = StoppedVM(uuid: "machine-uuid", name: "machine-name")
-        try vm.set(.sharedSmartcard(.off), runner: runner)
-        XCTAssertEqual(runner.command, "prlctl set machine-uuid --auto-share-smart-card off")
-    }
-
-    func testThatSetVMisolationToOnWorks() throws {
-        let runner = TestCommandRunner()
-        let vm = StoppedVM(uuid: "machine-uuid", name: "machine-name")
-        try vm.set(.isolateVM(.on), runner: runner)
-        XCTAssertEqual(runner.command, "prlctl set machine-uuid --isolate-vm on")
-    }
-
-    func testThatSetVMisolationToOffWorks() throws {
-        let runner = TestCommandRunner()
-        let vm = StoppedVM(uuid: "machine-uuid", name: "machine-name")
-        try vm.set(.isolateVM(.off), runner: runner)
-        XCTAssertEqual(runner.command, "prlctl set machine-uuid --isolate-vm off")
+        XCTAssertEqual(
+            runner.commands.last,
+            "prlctl snapshot-delete machine-uuid -i {64d481bb-ce04-45b1-8328-49e4e4c43ddf}"
+        )
     }
 
     func testThatVMEqualityIsBasedOnUUID() throws {
-        let vm1 = VM(uuid: "uuid", name: "name 1", status: .stopped, ip_configured: "127.0.0.1")
-        let vm2 = VM(uuid: "uuid", name: "name 2", status: .stopped, ip_configured: "127.0.0.1")
+        let vm1 = VM(uuid: "uuid", name: "name 1", status: .stopped, ipAddress: "127.0.0.1")
+        let vm2 = VM(uuid: "uuid", name: "name 2", status: .stopped, ipAddress: "127.0.0.1")
         XCTAssertEqual(vm1, vm2)
     }
 }
